homepage code
I've tried to make a start on the homepage header but it's not great. The idea was to have our logo on the right side and then directly opposite would be links to each of our pages but I haven't been able to match them up correctly. Then underneath the header is the image section which I've tried and failed to make cover the screen. Idea is to have sections underneath like mini desc of our pages, the newsletter sign up, social media links, copyright etc so when we go to coding our own pages we copy paste the homepage index code and just remove the image section and desc section gets transformed to our sections.